Move autoupdate code in proxy to make more sense
